+++ core_lib/string.h
@@ -10,7 +10,7 @@ void itos(uint32_t num, char *str);
 
 bool pos_strcmp(const char *str1, const char *str2, size_t size);
 void *pos_memcpy(char *dst, const char *src, size_t count);
-void *pos_memset(void *mem, int val, size_t count);
+void *pos_memset(void *mem, uint8_t val, size_t count);
 
 // Looks like good code (no)
 // this is hack to make core_lib compilable with stdlib
